requirements.txt
aiofiles==23.1.0
aiogram==2.25.1
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==4.0.0
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.7
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fastapi==0.103.0
frozenlist==1.4.0
h11==0.14.0
idna==3.4
magic-filter==1.0.10
multidict==6.0.4
openpyxl==3.1.2
pydantic==2.3.0
pydantic_core==2.6.3
python-dotenv==1.0.0
pytz==2023.3
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2
yarl==1.9.2
